Change url on filter change
