+++ package.json
@@ -48,7 +48,7 @@
     "react-dom": "16.13.1",
     "sass": "1.27.0",
     "swiper": "6.3.4",
-    "vanilla-framework": "2.20.0",
+    "vanilla-framework": "2.21.0",
     "watch-cli": "0.2.3",
     "webpack": "4.44.1",
     "webpack-cli": "3.3.12"
