+++ static/sass/_charmhub_l-store.scss
@@ -1,0 +1,160 @@
+@mixin l-charmhub-store {
+  .l-store {
+    @media screen and(min-width: $breakpoint-small) {
+      display: flex;
+      max-width: $grid-max-width;
+    }
+
+    @media screen and (max-width: $breakpoint-medium - 1) {
+      display: block;
+    }
+
+    .l-navigation {
+      @media screen and(min-width: $breakpoint-medium) {
+        margin-inline-end: 2rem;
+        min-width: 13.5rem;
+        width: 13.5rem;
+      }
+
+      @media screen and (min-width: $breakpoint-medium) {
+        display: block;
+      }
+
+      .p-side-navigation__drawer,
+      .p-side-navigation__drawer-header {
+        background-color: $color-light;
+      }
+
+      .p-side-navigation__drawer-header {
+        margin-block-end: 1rem;
+
+        @media screen and(min-width: $breakpoint-medium) {
+          margin-block-end: 0;
+        }
+      }
+
+      .p-muted-heading,
+      .p-side-navigation__list {
+        padding-inline-start: 1rem;
+
+        @media screen and(min-width: $breakpoint-medium) {
+          padding-inline-start: 0.5rem;
+        }
+      }
+
+      .p-side-navigation__list {
+        .p-checkbox {
+          margin-block-end: 0;
+        }
+      }
+    }
+
+    .l-main {
+      width: 100%;
+
+      .l-main__header {
+        align-items: center;
+        display: flex;
+
+        .l-main__header--left {
+          width: 25%;
+        }
+
+        .l-main__header--right {
+          display: flex;
+          justify-content: flex-end;
+          width: 75%;
+
+          .p-container--inline {
+            display: flex;
+            justify-content: space-between;
+
+            .p-button {
+              margin-inline-end: 1rem;
+            }
+
+            .p-button,
+            .p-form {
+              width: 50%;
+            }
+          }
+
+          @media screen and (max-width: $breakpoint-medium - 1) {
+            display: block;
+            width: 100%;
+          }
+        }
+
+        .p-search-box {
+          margin-inline-end: $spv-inner--large;
+          width: 100%;
+        }
+
+        @media screen and (max-width: $breakpoint-medium - 1) {
+          flex-direction: column-reverse;
+
+          .l-main__header--left {
+            display: flex;
+            justify-content: space-between;
+            width: 100%;
+          }
+
+          .l-main__header--right {
+            width: 100%;
+          }
+
+          .p-layout__search {
+            margin-inline-end: 0;
+          }
+
+          .p-layout__sort-desktop {
+            display: block;
+          }
+        }
+
+        @media screen and(min-width: 926px) and(max-width: 1150px) {
+          .l-main__header--left {
+            width: 33.33%;
+          }
+
+          .l-main__header--right {
+            width: 66.67%;
+          }
+        }
+      }
+
+      .l-flex__card-container {
+        display: flex;
+        flex-wrap: wrap;
+        margin-inline-end: -1rem;
+
+        .p-notification--information {
+          margin-inline-end: 1rem;
+          width: 100%;
+        }
+      }
+
+      .l-flex__card {
+        padding-block-end: 1rem;
+        padding-inline-end: 1rem;
+        width: 20%;
+
+        @media screen and(min-width: 1151px) and(max-width: 1600px) {
+          width: 25%;
+        }
+
+        @media screen and(min-width: 926px) and(max-width: 1150px) {
+          width: 33.33%;
+        }
+
+        @media screen and(min-width: $breakpoint-x-small - 1) and(max-width: 925px) {
+          width: 50%;
+        }
+
+        @media screen and(max-width: $breakpoint-x-small) {
+          width: 100%;
+        }
+      }
+    }
+  }
+}
