+++ templates/partial/_topic-cards.html
@@ -1,4 +1,4 @@
-<div class="p-layout__card u-equal-height" data-js="topic-card" data-filter="kubernetes">
+<div class="l-flex__card u-equal-height" data-js="topic-card" data-filter="kubernetes">
   <a href="/topics/kubernetes" class="p-card--button is-topic-card">
     <div class="p-card__header" style="background-image: linear-gradient(to bottom left, rgba(255, 255, 255, 0.1) 0, rgba(255, 255, 255, 0.1) 49.9%, transparent 50%), linear-gradient(to top left, rgba(255, 255, 255, 0.1) 0, rgba(255, 255, 255, 0.1) 49.9%, transparent 50%), linear-gradient(252deg, #014bd1 45%, #4279f4 91%);">
       <img src="https://assets.ubuntu.com/v1/6c605f4d-Group+27+Copy.svg" alt="" class="p-card__thumbnail">
