+++ templates/store-search.html
@@ -14,16 +14,16 @@
 </section>
 <div class="p-strip--accent u-no-padding--top u-no-padding--bottom">
   <section class="p-strip--main u-fixed-width is-wide">
-    <div class="p-layout--store">
-      <div class="p-layout__main">
-        <div class="p-layout__filter-container">
-          <div class="p-layout__left">
+    <div class="l-store">
+      <div class="l-main">
+        <div class="l-main__header">
+          <div class="l-main__header--left">
             <div class="p-muted-heading" id="results-count-container" data-js="filtered-items">{{ results|length }} result{% if results|length != 1 %}s{% endif %}</div>
           </div>
         </div>
-        <div class="p-layout__card-container" id="features-container">
+        <div class="l-flex__card-container" id="features-container">
           {% for charm in results %}
-            <div class="p-layout__card u-equal-height" id="mysql">
+            <div class="l-flex__card u-equal-height" id="mysql">
               <a href="/{{ charm['name'] }}" class="p-card--button">
                 <div class="p-card__header">
                   <div class="p-card__thumbnail-container">
@@ -49,7 +49,7 @@
                       <img alt="Kubernetes" src="https://assets.ubuntu.com/v1/f1852c07-Kubernetes.svg" width="24" height="24">
                       <span class="p-tooltip__message" role="tooltip">This operator drives the application on Kubernetes</span>
                     {% elif charm.store_front.os[0] == "linux" %}
-                      <div class="entity-card-icons"><img alt="Linux" src="https://assets.ubuntu.com/v1/dc11bd39-Linux.svg" width="24" height="24"></div>
+                      <img alt="Linux" src="https://assets.ubuntu.com/v1/dc11bd39-Linux.svg" width="24" height="24">
                       <span class="p-tooltip__message" role="tooltip">This operator drives the application on Linux servers</span>
                     {% endif %}
                     </span>
