+++ templates/store.html
@@ -12,38 +12,132 @@
     <p><a href="/manifesto" class="p-link--inverted">Read our manifesto&nbsp;&rsaquo;</a></p>
   </div>
 </section>
-<div class="p-strip--accent u-no-padding--top u-no-padding--bottom">
-  <div class="p-sort" data-js="mobile-platform-handler">
-    <div class="p-sort__title">
-      Platform
+<section class="p-strip--accent u-no-padding--top u-no-padding--bottom">
+  <div class="p-strip--main u-fixed-width is-wide">
+    <div class="l-store">
+      <div class="l-navigation">
+        <div class="p-side-navigation" id="drawer">
+          <div class="p-side-navigation__overlay" aria-controls="drawer"></div>
+          <nav class="p-side-navigation__drawer" aria-label="Filters drawer">
+            <div class="p-side-navigation__drawer-header">
+              <a href="#" class="p-side-navigation__toggle--in-drawer js-drawer-toggle" aria-controls="drawer">Hide filters</a>
+            </div>
+            <h3 class="p-muted-heading u-hide--small" data-filters="applied-filters">
+              Filters
+            </h3>
+            <form>
+              {% if categories %}
+              <ul class="p-side-navigation__list">
+                {% for category in categories %}
+                {% if category.slug != "featured" and category.slug != "other" %}
+                <li class="p-side-navigation__item" data-filter-type="category" data-js="filter">
+                  <label class="p-checkbox">
+                    <input class="p-checkbox__input category-filter" aria-labelledby="{{ category.slug }}-filter" type="checkbox" value="{{ category.slug }}" {% if active_filter('category', category.slug) %}checked{% endif %} disabled>
+                    <span class="p-checkbox__label" id="{{ category.slug }}-filter">{{ category.name }}</span>
+                  </label>
+                </li>
+                {% endif %}
+                {% endfor %}
+              </ul>
+              {% endif %}
+            </form>
+            <hr />
+            <h3 class="p-muted-heading">
+              Useful links
+            </h3>
+            <ul class="p-side-navigation__list">
+              <li class="p-side-navigation__item">
+                <a href="/about">What is an operator?</a>
+              </li>
+              <li class="p-side-navigation__item">
+                <a href="/publishing">How do I publish here?</a>
+              </li>
+              <li class="p-side-navigation__item">
+                <a href="/glossary">Glossary</a>
+              </li>
+            </ul>
+          </nav>
+        </div>
+      </div>
+      <div class="l-main">
+        <div class="l-main__header">
+          <div class="l-main__header--left">
+            <div class="p-muted-heading" id="results-count-container" data-js="filtered-items">{{ featured_charms|length }} Featured</div>
+          </div>
+          <div class="l-main__header--right">
+            <form action="/" class="p-search-box u-hide--medium u-hide--large" data-js="search-handler-mobile">
+              <input type="search" class="p-search-box__input" name="q" aria-label="Search Charmhub" placeholder="Search Charmhub" {% if q %}value="{{ q }}" {% endif %} required />
+              <button type="reset" class="p-search-box__reset"><i class="p-icon--close">Clear</i></button>
+              <button type="submit" class="p-search-box__button" alt="Search"><i class="p-icon--search">Search</i></button>
+            </form>
+            <div class="p-container--inline">
+              <!-- <a href="#drawer" class="p-side-navigation__toggle js-drawer-toggle" aria-controls="drawer">Filters (4)</a> -->
+              <a href="#drawer" class="p-button has-icon u-hide--medium u-hide--large js-drawer-toggle" data-js="filter-button-mobile-open"><i class="p-icon--arrow-right"></i><span>Filters</span></a>
+              <form class="p-form p-form--inline">
+                <div class="p-form__group u-no-margin--right">
+                  <label for="platform" aria-label="Filter platforms" class="p-form__label u-hide--small">Platforms</label>
+                  <select name="platform" id="platform-handler" data-js="platform-handler" class="p-form__control" value="{{ platform }}" id="platform" disabled>
+                    <option value="all" {% if active_filter('platform', 'all' ) %}selected{% endif %}>All</option>
+                    <option value="linux" {% if active_filter('platform', 'linux' ) %}selected{% endif %}>Linux</option>
+                    <option value="kubernetes" {% if active_filter('platform', 'kubernetes' ) %}selected{% endif %}>Kubernetes</option>
+                  </select>
+                </div>
+              </form>
+            </div>
+          </div>
+        </div>
+
+        <div class="l-flex__card-container" id="features-container">
+          {% include "partial/_featured-charms.html" %}
+        </div>
+
+        <div class="l-flex__card-container u-hide" id="entity-container">
+          <div><i class="p-icon--spinner u-animation--spin"></i> Loading&hellip;</div>
+        </div>
+
+        <div class="u-align--center" style="margin-top: 1rem;">
+          <button type="button" class="p-button--positive" id="show-all-charms" disabled>See all operators</button>
+        </div>
+      </div>
     </div>
-    <form class="p-sort__content">
-      <div class="p-sort__item">
-        <input type="radio" name="platformMobile" id="all" value="all" {% if active_filter('platform', 'all') %}checked{% endif %}>
-        <label for="all">All</label>
-      </div>
-      <div class="p-sort__item">
-        <input type="radio" name="platformMobile" id="linux" value="linux" {% if active_filter('platform', 'linux') %}checked{% endif %}>
-        <label for="linux">Linux</label>
-      </div>
-      <div class="p-sort__item">
-        <input type="radio" name="platformMobile" id="windows" value="windows" {% if active_filter('platform', 'windows') %}checked{% endif %}>
-        <label for="windows">Windows</label>
-      </div>
-      <div class="p-sort__item">
-        <input type="radio" name="platformMobile" id="kubernetes" value="kubernetes" {% if active_filter('platform', 'kubernetes') %}checked{% endif %}>
-        <label for="kubernetes">Kubernetes</label>
-      </div>
-    </form>
   </div>
-  <section class="p-strip--main u-fixed-width is-wide">
-    <div class="p-layout--store">
+</section>
 
-      {% include "partial/_filters.html" %}
+<template id="entity-card">
+  <div class="l-flex__card u-equal-height">
+    <a href="" class="p-card--button">
+      <div class="p-card__header">
+        <div class="p-card__thumbnail-container">
+          <img src="" alt="" width="40" height="40" class="p-card__thumbnail">
+        </div>
+        <h3 class="p-card__title p-heading--4 u-no-margin--bottom entity-card-title"></h3>
+        <p class="u-text--muted u-no-padding--top entity-card-publisher"></p>
+      </div>
 
-  </section>
-</div>
+      <div class="p-card__content">
+        <p><small class="entity-card-summary"></small></p>
+      </div>
 
+      <div class="p-card__footer">
+        <div class="entity-card-icons"></div>
+      </div>
+    </a>
+  </div>
+</template>
+
+<style>
+  .entity-card-title {
+    text-transform: capitalize;
+  }
+
+  .entity-card-summary {
+    display: inline-block;
+  }
+
+  .entity-card-summary::first-letter {
+    text-transform: uppercase;
+  }
+</style>
 {% endblock %}
 
 {% block page_scripts %}
